fix(plugin): rename module to avoid having same name as containing module
